MAPSJS-2660: Move TileOffsetUtils to TileKeyUtils in harp-geoutils.
